+++ app/src/main/java/com/model/ActionID.java
@@ -1,9 +1,9 @@
 package com.model;
 
 public class ActionID {
-    public final static int move = 1001;
-    public final static int attack1 = 1002;
-    public final static int attack2 = 1003;
-    public final static int back = 1201;
-    public final static int nextTurn = 1200;
+    public final static int move = (int) 1001;
+    public final static int attack1 = (int) 1002;
+    public final static int attack2 = (int) 1003;
+    public final static int back = (int) 1201;
+    public final static int nextTurn = (int) 1200;
 }
